+++ db/migrate/20191121132035_add_subject_to_formal_performance_action_to_maths_and_physics_eligibilities.rb
@@ -1,0 +1,5 @@
+class AddSubjectToFormalPerformanceActionToMathsAndPhysicsEligibilities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :maths_and_physics_eligibilities, :subject_to_formal_performance_action, :boolean
+  end
+end